a23_send_email/requirements.txt
beautifulsoup4==4.9.3
lxml==4.8.0
selenium==3.141.0
redis==3.5.3
discord.py==1.7.3